+++ src/resources/js/Components/app/GroupList.vue
@@ -1,6 +1,6 @@
 <script setup>
-import GroupItem from '@/Components/app/GroupItem.vue';
 import TextInput from '@/Components/TextInput.vue';
+import GroupItem from '@/Components/app/GroupItem.vue';
 import {ref} from "vue";
 
 const searchKeyword = ref('')