+++ src/resources/js/Pages/Home.vue
@@ -1,6 +1,7 @@
 <script setup>
 import { Head, Link } from '@inertiajs/vue3';
 import GroupList from "@/Components/app/GroupList.vue";
+import FollowingList from "@/Components/app/FollowingList.vue";
 
 defineProps({
 
@@ -18,7 +19,7 @@ defineProps({
             Posts
         </div>
         <div class="col-span-3">
-            Followers
+            <FollowingList />
         </div>
 
     </div>