+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     chown 1000:1000 app/Http/Controllers/HomeController.php
     ```
     - created the dashboard page as simple html, convert parts of it to component
+    - added `Following` list as components, like in `My Groups`
 
 
 - 2024-02-18